Fix adding courses
